--- src/main/java/com/example/koun/Repository/ItemRepository.java
@@ -1,10 +1,0 @@
-package com.example.koun.Repository;
-
-import com.example.koun.domain.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository <Item,Long> {
-
-}
