--- src/main/java/com/example/koun/Repository/OrderRepository.java
@@ -1,11 +1,0 @@
-package com.example.koun.Repository;
-
-import com.example.koun.domain.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-
-}
