--- src/main/java/com/example/koun/Repository/RaffleRepository.java
@@ -1,8 +1,0 @@
-package com.example.koun.Repository;
-
-import com.example.koun.domain.Raffle;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RaffleRepository extends JpaRepository <Raffle, Long> {
-
-}
