--- src/main/java/com/example/koun/Repository/SectionRepository.java
@@ -1,8 +1,0 @@
-package com.example.koun.Repository;
-
-import com.example.koun.domain.Section;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepository extends JpaRepository <Section, Long> {
-
-}
