--- src/main/java/com/example/koun/Repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.koun.Repository;
-
-import com.example.koun.domain.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<Item, Long> {
-
-
-}
