+++ src/main/java/com/example/koun/repository/UserRepository.java
@@ -1,8 +1,7 @@
-package com.example.koun.Repository;
+package com.example.koun.repository;
 
 
 import com.example.koun.domain.User;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
