+++ src/main/java/com/example/koun/service/OrderService.java
@@ -1,4 +1,4 @@
-package com.example.koun.Service;
+package com.example.koun.service;
 
 public interface OrderService {
